django_project/requirements/requirements.txt
Django
django-ckeditor-5
django-htmx
django-resized
django-robots
django-simple-captcha
feedparser
filetype
Pillow
psycopg[binary]
psycopg
psutil
python-dotenv
readtime
requests


# For OpenAI
scipy
numpy
plotly
openai
pandas
scikit-learn
matplotlib
tiktoken

# Legacy dependencies for migrations
django-ckeditor==6.4.2

# Development dependencies
django-coverage-plugin
coverage
black
django-fastdev
pytest
pre-commit
django-livereload-server
ruff

# Production dependencies
gunicorn
whitenoise
django-sri
django-csp
Brotli